+++ src/components/Section.tsx
@@ -1,0 +1,24 @@
+import {
+  ListItem,
+  ListItemButton,
+  ListItemIcon,
+  ListItemText,
+} from "@mui/material";
+import InboxIcon from "@mui/icons-material/MoveToInbox";
+import MailIcon from "@mui/icons-material/Mail";
+import { ISection } from "../interfaces/layout-interfaces";
+
+const Section = ({ text, index }: ISection) => {
+  return (
+    <ListItem key={text} disablePadding>
+      <ListItemButton>
+        <ListItemIcon>
+          {index % 2 === 0 ? <InboxIcon /> : <MailIcon />}
+        </ListItemIcon>
+        <ListItemText primary={text} />
+      </ListItemButton>
+    </ListItem>
+  );
+};
+
+export default Section;
