+++ src/interfaces/layout-interfaces.ts
@@ -1,0 +1,4 @@
+export interface ISection {
+  text: string;
+  index: number;
+}
